Replace original cpsat solvers by cpsat-auto versions
For:
- rcpsp
- rcpsp_multiskill
- workforce/scheduling
