+++ examples/rcpsp/run_cpsat_from_partial_solution.py
@@ -9,7 +9,7 @@ from discrete_optimization.generic_tasks_tools.solvers.cpsat.auto_impl import (
 )
 from discrete_optimization.rcpsp.parser import get_data_available, parse_file
 from discrete_optimization.rcpsp.solution import RcpspSolution
-from discrete_optimization.rcpsp.solvers.cpsat_auto import CpSatAutoRcpspSolver
+from discrete_optimization.rcpsp.solvers.cpsat import CpSatRcpspSolver
 from discrete_optimization.rcpsp.transformations.generic_scheduling_impl import (
     GenericSchedulingToRcpspTransformation,
     RcpspToGenericSchedulingTransformation,
@@ -58,7 +58,7 @@ back_transformation = GenericSchedulingToRcpspTransformation()
 rcpsp_subproblem = back_transformation.transform_problem(
     source_problem=generic_subproblem
 )
-rcpsp_solver = CpSatAutoRcpspSolver(problem=rcpsp_subproblem)
+rcpsp_solver = CpSatRcpspSolver(problem=rcpsp_subproblem)
 rcpsp_subsolution: RcpspSolution = rcpsp_solver.solve(time_limit=10).get_best_solution()
 # reconstructing solution is easier generic side
 generic_full_solution = GenericSchedulingImplSolution(
