+++ tests/rcpsp/solvers/test_cpsat.py
@@ -87,12 +87,13 @@ def test_ortools(model):
     "model",
     ["j301_1.sm", "j1010_1.mm"],
 )
-def test_objectives(model):
+@pytest.mark.parametrize("avoid_interval_optional", [False, True])
+def test_objectives(model, avoid_interval_optional):
     files_available = get_data_available()
     file = [f for f in files_available if model in f][0]
     rcpsp_problem = parse_file(file)
     solver = CpSatRcpspSolver(problem=rcpsp_problem)
-    solver.init_model()
+    solver.init_model(avoid_interval_optional=avoid_interval_optional)
 
     subtasks = {1, 4}
     # max end time subtasks
@@ -242,7 +243,8 @@ def test_chaining_constraints():
     "model",
     ["j301_1.sm", "j1010_1.mm"],
 )
-def test_ortools_with_calendar_resource(model):
+@pytest.mark.parametrize("avoid_interval_optional", [False, True])
+def test_ortools_with_calendar_resource(model, avoid_interval_optional):
     files_available = get_data_available()
     file = [f for f in files_available if model in f][0]
     rcpsp_problem = parse_file(file)
@@ -254,15 +256,18 @@ def test_ortools_with_calendar_resource(model):
     rcpsp_problem.update_problem()
     assert rcpsp_problem.is_calendar
     solver = CpSatRcpspSolver(problem=rcpsp_problem)
-    result_storage = solver.solve(time_limit=100)
+    result_storage = solver.solve(
+        time_limit=100, avoid_interval_optional=avoid_interval_optional
+    )
     solution, fit = result_storage.get_best_solution_fit()
     solution_rebuilt = RcpspSolution(
         problem=rcpsp_problem,
         rcpsp_permutation=solution.rcpsp_permutation,
         rcpsp_modes=solution.rcpsp_modes,
     )
-    fit_2 = rcpsp_problem.evaluate(solution_rebuilt)
-    assert fit == -fit_2["makespan"]
+    eval_dict = rcpsp_problem.evaluate(solution_rebuilt)
+    print(eval_dict)
+    assert fit == -eval_dict["makespan"]
     assert rcpsp_problem.satisfy(solution)
     assert solution.check_all_calendar_resource_capacity_constraints()
 
@@ -371,9 +376,10 @@ def test_special_constraints():
 
     resources = {"R1": 2}
 
-    # Test 1: start_together constraint
+    # Test 1: start_together constraint + start times
     special_constraints_1 = SpecialConstraintsDescription(
         start_together=[(2, 3)],  # tasks 2 and 3 start together
+        start_times={4: 10},  # task 4 should start at 10
     )
 
     problem_1 = RcpspProblem(
@@ -395,11 +401,13 @@ def test_special_constraints():
         f"task 2 starts at {solution_1.get_start_time(2)}, "
         f"task 3 starts at {solution_1.get_start_time(3)}"
     )
+    assert solution_1.get_start_time(4) == 10
     assert problem_1.satisfy(solution_1), "Solution should satisfy all constraints"
 
-    # Test 2: start_at_end constraint
+    # Test 2: start_at_end constraint + end times
     special_constraints_2 = SpecialConstraintsDescription(
         start_at_end=[(3, 4)],  # task 4 starts when task 3 ends
+        end_times={5: 20},  # task 5 ends at 20
     )
 
     problem_2 = RcpspProblem(
@@ -421,6 +429,7 @@ def test_special_constraints():
         f"task 3 ends at {solution_2.get_end_time(3)}, "
         f"task 4 starts at {solution_2.get_start_time(4)}"
     )
+    assert solution_2.get_end_time(5) == 20
     assert problem_2.satisfy(solution_2), "Solution should satisfy all constraints"
 
     # Test 3: start_times_window constraint
